Extract common resource provider management.
